requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.12.2
billiard==4.2.1
bleach==6.2.0
browsermob-proxy==0.8.0
celery==5.3.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-driver==0.2.5
cron-descriptor==1.4.5
cryptography==44.0.0
defusedxml==0.7.1
diff-match-patch==20241021
distro==1.9.0
Django==5.1.7
django-admin-inline-paginator==0.4.0
django-admin-interface==0.29.4
django-admin-material-dashboard==1.0.10
django-allauth==0.57.0
django-appconf==1.0.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cleanup==9.0.0
django-clickhouse==1.2.2
django-colorfield==0.12.0
django-compressor==4.3.1
django-cors-headers==4.3.1
django-debug-toolbar==4.4.6
django-environ==0.8.1
django-extensions==3.2.3
django-extra-views==0.14.0
django-filter==24.3
django-haystack==3.3.0
django-imagekit==5.0.0
django-import-export==4.3.7
django-jazzmin==2.6.0
django-modeltranslation==0.19.13
django-oscar==3.2.5
django-phonenumber-field==6.4.0
django-redis==5.2.0
django-simpleui==2025.1.13
django-storages==1.14.5
django-summernote==0.8.20.0
django-tables2==2.3.4
django-timezone-field==7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dnspython==2.7.0
drf-nested-routers==0.94.1
drf-yasg==1.21.10
elastic-transport==8.17.0
elasticsearch==7.17.12
et_xmlfile==2.0.0
factory-boy==3.2.1
fake-useragent==2.1.0
Faker==37.0.0
fastapi==0.104.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
infi.clickhouse-orm==2.1.3
inflection==0.5.1
iso8601==2.1.0
Js2Py==0.74
kombu==5.4.2
livereload==2.7.1
loguru==0.7.2
lxml==4.9.3
m3u8==3.6.0
msgpack==1.1.0
mysql-connector-python==9.2.0
numpy==2.2.4
oauthlib==3.2.2
openai==1.6.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
phonenumbers==9.0.0
pilkit==3.0
pillow==11.0.0
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2-binary==2.9.9
purl==1.6
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.5.2
pydantic_core==2.14.5
pyjsparser==2.7.1
PyJWT==2.9.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==3.2.1
python-magic==0.4.27
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rjsmin==1.2.1
schedule==1.2.2
selenium==4.15.2
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
starlette==0.27.0
statsd==4.0.1
supervisor==4.2.4
tablib==3.8.0
text-unidecode==1.3
tornado==6.4.2
tqdm==4.66.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.24.0
uWSGI==2.0.28
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==12.0
whitenoise==6.9.0
wsproto==1.2.0
